Fix mutation inside of flat assemblages

+++ RZAssemblage/RZFRCAssemblage.m
@@ -10,7 +10,7 @@
 #import "NSIndexPath+RZAssemblage.h"
 #import "RZAssemblage+Private.h"
 
-#define RZAssertIndexPathLength(indexPath, offset) RZRaize(indexPath.length <= ((self.hasSections ? 2 : 1) + offset), @"Index Path %@ has length of %lu, expected index <= %d", indexPath, indexPath.length, ((self.hasSections ? 2 : 1) + offset))
+#define RZAssertIndexPathLength(indexPath, offset) RZRaize(indexPath.length <= ((self.hasSections ? 2 : 1) + offset), @"Index Path %@ has length of %lu, expected index <= %d", indexPath, (unsigned long)indexPath.length, ((self.hasSections ? 2 : 1) + offset))
 #define RZAssertContainerIndexPath(indexPath) RZAssertIndexPathLength(indexPath, -1)
 #define RZAssertItemIndexPath(indexPath) RZAssertIndexPathLength(indexPath, 0)
